Mostly finished level editor and cleaned up
The level editor is now usable, file saving and loading is possible. The new levels can be used in game now and are randomly chosen from a map list which is automatically updated when you make a new level/map. Cleaned up some code to be more in line with my practices.